% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.corrmatrix.R
\name{is.corrmatrix}
\alias{is.corrmatrix}
\title{Correlation matrix}
\usage{
is.corrmatrix(matrix)
}
\arguments{
\item{matrix}{a (non-empty) numeric matrix of data values.}
}
\value{
A logical value: True/False.
}
\description{
Checks if a given matrix is a correlation matrix for non-degenerate distributions.
}
\examples{

m1<-matrix(c(1,2,2,1),nrow = 2,byrow = TRUE)
is.corrmatrix(m1)

m2<-matrix(c(1,0.8,0.8,1),nrow = 2,byrow = TRUE)
is.corrmatrix(m2)

m3<-matrix(c(1,0.7,0.8,1),nrow = 2,byrow = TRUE)
is.corrmatrix(m3)

}
