% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teachingstatdata.R
\docType{package}
\name{stats4teaching}
\alias{stats4teaching}
\title{Teaching Statistics Data Simulation}
\description{
Univariate and multivariate normal data simulation. They also supply a brief summary of the analysis for each experiment/design.
\itemize{

 \item Independent samples.

 \item One-way and two-way ANOVA.

 \item Paired samples (T-Test & Regression).

 \item Repeated measures (ANOVA & Multiple Regression).

 \item Clinical Assay.

}
}
\author{
Esteban Cabello García and Pedro Jesús Femia Marzo.
}
